+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0](https://github.com/nyurik/fast-mvt/compare/v0.4.1...v0.5.0) - 2026-07-12
+
+### Other
+
+- debug format support, cleaner api ([#26](https://github.com/nyurik/fast-mvt/pull/26))
+- minor internal cleanup ([#23](https://github.com/nyurik/fast-mvt/pull/23))
+
 ## [0.4.1](https://github.com/nyurik/fast-mvt/compare/v0.4.0...v0.4.1) - 2026-06-23
 
 ### Other